refactor: update ContentType imports from programs to section-contents across multiple files
