+++ src/components/public/detailed-program-section/DetailedProgramSection.test.tsx
@@ -1,12 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import { useLocation } from 'react-router-dom';
 import { DetailedProgramSection } from './DetailedProgramSection';
-import {
-    HippotherapyProgramSectionDto,
-    SectionTemplate,
-    SectionMode,
-} from '@/types/common/program-sections';
-import { ContentType } from '@/types/common/programs';
+import { HippotherapyProgramSectionDto, SectionTemplate, SectionMode } from '@/types/common/program-sections';
+import { ContentType } from '@/types/common/section-contents';
 import * as renderProgramSectionModule from '@/utils/functions/render-program-section';
 
 jest.mock('@/utils/functions/render-program-section', () => ({
