+++ src/pages/admin/programs/components/translate-program-form/TranslateProgramForm.test.tsx
@@ -6,7 +6,7 @@ import { TranslateProgramForm, TranslateProgramFormRef } from './TranslateProgra
 import { PROGRAMS_TEXT } from '@/const/admin/programs';
 import { VisibilityStatus } from '@/types/admin/common';
 import { CreateHippotherapyProgramSectionLocalizationDto } from '@/types/common/program-sections';
-import { ContentType } from '@/types/common/programs';
+import { ContentType } from '@/types/common/section-contents';
 import { SectionTemplate } from '@/types/common/program-sections';
 
 const createFormControlMock = (element: 'input' | 'textarea') => {
